Bump version to 2.0, require libcups v3.

+++ pappl/mainloop.h
@@ -1,7 +1,7 @@
 //
 // Mainloop header file for the Printer Application Framework
 //
-// Copyright © 2020-2022 by Michael R Sweet.
+// Copyright © 2020-2023 by Michael R Sweet.
 //
 // Licensed under Apache License v2.0.  See the file "LICENSE" for more
 // information.
@@ -20,9 +20,9 @@ extern "C" {
 // Types...
 //
 
-typedef int (*pappl_ml_subcmd_cb_t)(const char *base_name, int num_options, cups_option_t *options, int num_files, char **files, void *data);
+typedef int (*pappl_ml_subcmd_cb_t)(const char *base_name, size_t num_options, cups_option_t *options, size_t num_files, char **files, void *data);
 					// Sub-command callback
-typedef pappl_system_t *(*pappl_ml_system_cb_t)(int num_options, cups_option_t *options, void *data);
+typedef pappl_system_t *(*pappl_ml_system_cb_t)(size_t num_options, cups_option_t *options, void *data);
 					// System callback
 typedef void (*pappl_ml_usage_cb_t)(void *data);
 					// Program usage callback
@@ -32,7 +32,7 @@ typedef void (*pappl_ml_usage_cb_t)(void *data);
 // Functions...
 //
 
-extern int	papplMainloop(int argc, char *argv[], const char *version, const char *footer_html, int num_drivers, pappl_pr_driver_t *drivers, pappl_pr_autoadd_cb_t autoadd_cb, pappl_pr_driver_cb_t driver_cb, const char *subcmd_name, pappl_ml_subcmd_cb_t subcmd_cb, pappl_ml_system_cb_t system_cb, pappl_ml_usage_cb_t usage_cb, void *data) _PAPPL_PUBLIC;
+extern int	papplMainloop(int argc, char *argv[], const char *version, const char *footer_html, size_t num_drivers, pappl_pr_driver_t *drivers, pappl_pr_autoadd_cb_t autoadd_cb, pappl_pr_driver_cb_t driver_cb, const char *subcmd_name, pappl_ml_subcmd_cb_t subcmd_cb, pappl_ml_system_cb_t system_cb, pappl_ml_usage_cb_t usage_cb, void *data) _PAPPL_PUBLIC;
 extern void	papplMainloopShutdown(void) _PAPPL_PUBLIC;
